requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
fastapi==0.103.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.2
postgres==4.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pydantic==2.3.0
pydantic_core==2.6.3
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
simplejson==3.19.2
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.27.0
swagger-spec-validator==3.0.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5